removing module and integrate it in setup
